convae/requirements.txt
torch
numpy
gensim
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
tqdm==4.60.0
